requirements.txt
aiofiles==24.1.0
aioredis==2.0.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
fastapi==0.115.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jsonpickle==3.3.0
Mako==1.3.5
MarkupSafe==3.0.2
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMySQL==1.1.1
pytest==8.3.3
python-decouple==3.8
python-dotenv==1.0.1
redis==5.2.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.32.0
